+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- Fix `.update` and `.update!`: include given `nil` valued attributes in request (https://github.com/Beyond-Finance/active_force/pull/79)
 - Change `.first` to not query the API if records have already been retrieved (https://github.com/Beyond-Finance/active_force/pull/77)
 
 ## 0.20.1